# frontend/src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Food Recommender API",
...
        "contact": {
            "name": "Food Recommender Team",
            "email": "support@foodrecommender.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        },
        {
            "url": "https://api.foodrecommender.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/auth/google": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth Authentication",
                "description": "Authenticate or register a user using Google OAuth ID token",
                "operationId": "google_auth_auth_google_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GoogleAuthRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                },
                                "example": {
                                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                                    "token_type": "bearer",
                                    "user": {
                                        "id": "uuid-string",
                                        "email": "user@gmail.com",
                                        "first_name": "John",
                                        "last_name": "Doe",
                                        "google_id": "google_user_123",
                                        "profile_picture": "https://lh3.googleusercontent.com/...",
                                        "created_at": "2024-01-01T00:00:00",
                                        "is_active": true
                                    },
                                    "is_new_user": false
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid Google ID token",
                        "content": {
                            "application/json": {
                                "example": {
                                    "detail": "Invalid Google ID token: Token verification failed"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Get Current User",
                "description": "Get information about the currently authenticated user",
                "operationId": "get_current_user_info_auth_me_get",
                "responses": {
                    "200": {
                        "description": "User information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                },
                                "example": {
                                    "id": "uuid-string",
                                    "email": "user@gmail.com",
                                    "first_name": "John",
                                    "last_name": "Doe",
                                    "google_id": "google_user_123",
                                    "profile_picture": "https://lh3.googleusercontent.com/...",
                                    "created_at": "2024-01-01T00:00:00",
                                    "is_active": true
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "content": {
                            "application/json": {
                                "example": {
                                    "detail": "Could not validate credentials"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GoogleAuthRequest": {
                "properties": {
                    "id_token": {
                        "type": "string",
                        "title": "Id Token"
                    }
                },
                "type": "object",
                "required": [
                    "id_token"
                ],
                "title": "GoogleAuthRequest"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "LoginResponse": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "default": "bearer"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    },
                    "is_new_user": {
                        "type": "boolean",
                        "title": "Is New User",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "access_token",
                    "user"
                ],
                "title": "LoginResponse"
            },
...
            "UserResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "google_id": {
                        "type": "string",
                        "title": "Google Id"
                    },
                    "profile_picture": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Profile Picture"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "first_name",
                    "last_name",
                    "google_id",
                    "created_at",
                    "is_active"
                ],
                "title": "UserResponse"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}